# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ellowring API
  description: |
    Ellowring Software Solutions — Enterprise SaaS REST API (Phase 5).
    Learn. Prepare. Build. Get Hired.
  version: 1.0.0
  contact:
    name: Ellowring Engineering
    email: engineering@ellowring.com
servers:
  - url: http://localhost:4000/api
    description: Local
  - url: https://api-staging.ellowring.com/api
    description: Staging
  - url: https://api.ellowring.com/api
    description: Production
...
paths:
...
  /v1/partners/me/dashboard:
    get:
      tags:
        - Channel Partner
      summary: Partner dashboard KPIs.
      operationId: get__partners_me_dashboard
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/partners/me/leads:
    get:
      tags:
        - Channel Partner
      summary: Partner leads.
      operationId: get__partners_me_leads
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
    post:
      tags:
        - Channel Partner
      summary: Create lead.
      operationId: post__partners_me_leads
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
...
  /v1/partners/me/referrals:
    post:
      tags:
        - Channel Partner
      summary: Register referral attribution.
      operationId: post__partners_me_referrals
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/partners/me/wallet:
    get:
      tags:
        - Channel Partner
      summary: Commission wallet.
      operationId: get__partners_me_wallet
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/partners/me/commissions:
    get:
      tags:
        - Channel Partner
      summary: Commission ledger.
      operationId: get__partners_me_commissions
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/partners/me/payouts:
    post:
      tags:
        - Channel Partner
      summary: Request payout.
      operationId: post__partners_me_payouts
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
    get:
      tags:
        - Channel Partner
      summary: Payout history.
      operationId: get__partners_me_payouts
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
...
  /v1/payments/orders:
    post:
      tags:
        - Payment
      summary: Create payment order with payment provider (Razorpay).
      operationId: post__payments_orders
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/payments/verify:
    post:
      tags:
        - Payment
      summary: Verify provider payment signature and mark SUCCESS.
      operationId: post__payments_verify
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/payments/{id}/refunds:
    post:
      tags:
        - Payment
      summary: Refund a successful payment (full/partial).
      operationId: post__payments__id__refunds
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/invoices/{id}:
    get:
      tags:
        - Payment
      summary: Get invoice by id or payment.
      operationId: get__invoices__id_
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/premium-plans:
    get:
      tags:
        - Payment
      summary: List premium plans.
      operationId: get__premium-plans
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/subscriptions:
    post:
      tags:
        - Payment
      summary: Create/subscribe user to premium plan.
      operationId: post__subscriptions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/payments/transactions:
    get:
      tags:
        - Payment
      summary: Transaction history for current user.
      operationId: get__payments_transactions
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: 1
          schema:
            type: string
        - name: pageSize
          in: query
          required: false
          description: 20
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: "string?"
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
  /v1/payments/{id}:
    get:
      tags:
        - Payment
      summary: Get payment by id.
      operationId: get__payments__id_
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '429':
          description: Rate limited
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          example: true
        data: {}
        meta:
          type: object
          properties:
            requestId:
              type: string
            timestamp:
              type: string
              format: date-time
            pagination:
              $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        totalItems: { type: integer }
        totalPages: { type: integer }
        hasNext: { type: boolean }
        hasPrev: { type: boolean }